# Extend all analytic functions in base for quaternions