--- a/src/Quaternion.jl
+++ b/src/Quaternion.jl
@@ -178,7 +178,7 @@ function log(q::Quaternion)
 end
 
 (^)(q::Quaternion{T}, p::T) where {T<:Real} = extend_analytic(Base.Fix2(^, p), q)
-(^)(q::Quaternion{T}, p::S) where {T<:Real,S<:Real} = extend_analytic(Base.Fix2(^, p), q)
+(^)(q::Quaternion{T}, p::AbstractFloat) where {T<:Real} = extend_analytic(Base.Fix2(^, p), q)
 (^)(q::Quaternion, w::Quaternion) = exp(w * log(q))
 
 function linpol(p::Quaternion, q::Quaternion, t::Real)