Fixes #55. Also fixes #39 as a bonus.