@sethaxen 2022-02-20
For reference, this is the heuristic function I used to identify potentially complex analytic functions in Base:

```julia
using FiniteDifferences

fdm = central_fdm(5, 1)

function get_complex_analytic_funs(mod::Module, z=rand(ComplexF64); ignore_defaults=true)
    fnames = names(mod)
    return filter(fnames) do fname
        f = getproperty(mod, fname)
        f isa Union{DataType,UnionAll} && return false
        hasmethod(f, Tuple{ComplexF64}) || return false
        ignore_defaults && hasmethod(f, Tuple{Number}) && return false
        try
            f(z) isa Complex || return false
        catch e
            println("$f errored on input $z")
            return false
        end
        return is_analytic_at(f, z)
    end
end

function is_analytic_at(f, z)
    J = only(jacobian(fdm, f, z))
    # Cauchy–Riemann equations
    return J[1,1] ≈ J[2,2] && J[1, 2] ≈ -J[2, 1]
end
```

```julia
julia> get_complex_analytic_funs(Base) |> sort
27-element Vector{Symbol}:
 :acos
 :acosh
 :asin
 :asinh
 :atan
 :atanh
 :cis
 :cispi
 :cos
 :cosh
 :cospi
 :exp
 :exp10
 :exp2
 :expm1
 :log
 :log10
 :log1p
 :log2
 :round
 :sin
 :sinh
 :sinpi
 :sqrt
 :tan
 :tanh
 :√
```

`round` should be handled outside of this PR. The rest are extended here.

Of these, `cis` and `cispi` are the two that might be confusing for users and could be removed. Because this package chooses a particular embedding of the complex numbers in the quaternions, `exp(im*q)` is computeable, but this is not the same thing as the extension of `cis(z)` to the quaternions. In fact, the extension has an intimate relationship with the axis and angle of a rotation in the same way that `cis(z)` is related to a rotation.