@hyrodium 2022-02-22
The `exp(::Quaternion)` is now much faster than current master branch.

**Current master**
```julia
julia> @benchmark exp(Quaternion(0,1,2,3))
BenchmarkTools.Trial: 10000 samples with 991 evaluations.
 Range (min … max):  41.754 ns … 75.420 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     43.341 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   43.569 ns ±  2.194 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▆▂▁▁█▅                                                    ▁
  ▆▇██████▅▆▆▅▇▆▇▇▆▇▇▇▇▇▆▆▆▆▆▅▆▅▅▃▄▄▄▅▄▃▄▃▃▄▃▃▃▄▄▃▃▂▄▃▃▃▃▃▄▄▄ █
  41.8 ns      Histogram: log(frequency) by time      55.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

```

**With this PR**
```julia
julia> @benchmark exp(Quaternion(0,1,2,3))
BenchmarkTools.Trial: 10000 samples with 999 evaluations.
 Range (min … max):  11.564 ns … 350.632 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     11.894 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.043 ns ±   5.150 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                    █   ▇▇   █    ▂                             
  ▂▂▂▂▂▂▁▁▂▂▂▃▄▅▅▃▅██▅▆▄██▅▄▃█▄▃▂▃█▃▂▂▃▃▁▂▂▂▁▂▁▁▁▁▁▁▁▁▁▂▁▁▂▂▂▂ ▃
  11.6 ns         Histogram: frequency by time         12.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```