@sethaxen 2022-02-22
> The `exp(::Quaternion)` is now much faster than current master branch 🚀

It is faster, but not quite that much. You need to declare the quaternion before the benchmark or wrap it in a dollar sign, otherwise the compiler has too much information and cheats.

```julia
julia> q = Quaternion(0,1,2,3)
Quaternion{Int64}(0, 1, 2, 3, false)

julia> @btime exp($q)  # master
  53.776 ns (0 allocations: 0 bytes)
QuaternionF64(-0.8252990620752587, -0.15092132721996449, -0.30184265443992897, -0.45276398165989346, true)

julia> @btime exp($q)  # this PR
  39.218 ns (0 allocations: 0 bytes)
QuaternionF64(-0.8252990620752587, -0.15092132721996449, -0.30184265443992897, -0.45276398165989346, true)
```